Remove setting error_reporting
When using the Graph class error_reporting and display_errors get overwritten to E_ALL and true, which is very undesirable, especially for production. Removing this should change nothing when using JPGraph and only lead to a better experience. I suspect this was included to find a bug originally but then forgotten.