Theme auto update color and fill_color for barplot
If we set the color and fill_color for multiple barplots to draw an accumulated bar plot; The Graph->add function automatically updates the color and fill_color for the plots. We have to set color and fill_color again after calling Graph->add function.
In the above update i am checking if the color and fill_color are set or not.